+++ Picasso/Modules/SceneManager/SceneInterstitial.m
@@ -50,7 +50,7 @@
         self.description = sceneModel.description;
 
 		DataManager *dataManager = [DataManager sharedInstance];
-        int currentSceneId = [[dataManager getGameModel] currentScene];
+        NSInteger currentSceneId = [[dataManager getGameModel] currentScene];
 
         if(currentSceneId < [dataManager getScenesNumber] - 1) {
 	    	[[dataManager getSceneWithNumber:currentSceneId + 1] unlockScene];
